problem-service/requirements.txt
html2image==2.0.1
grpcio==1.49.1
grpcio-tools==1.49.1
protobuf==4.21.3